# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateUserRequest:
    properties:
      displayName:
        type: string
      email:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.SystemStats:
    properties:
      activeUsers:
        description: Users with activity in last 30 days
        type: integer
      totalFiles:
        type: integer
      totalSize:
        type: integer
      totalUsers:
        type: integer
      totalWorkspaces:
        type: integer
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      displayName:
        type: string
      email:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  handlers.WorkspaceStats:
    properties:
      totalFiles:
        type: integer
      totalSize:
        type: integer
      userEmail:
        type: string
      userID:
        type: integer
      workspaceCreatedAt:
        type: string
      workspaceID:
        type: integer
      workspaceName:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        minimum: 1
        type: integer
      lastWorkspaceId:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        enum:
        - admin
        - editor
        - viewer
    required:
    - email
    - id
    - role
    type: object
  models.UserRole:
    enum:
    - admin
    - editor
    - viewer
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleEditor
    - RoleViewer
...
info:
  contact: {}
  description: This is the API for Lemma markdown note taking app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Lemma API
  version: "1.0"
paths:
  /admin/stats:
    get:
      description: Get system-wide statistics as an admin
      operationId: adminGetSystemStats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SystemStats'
        "500":
          description: Failed to get file stats
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get system statistics
      tags:
      - Admin
  /admin/users:
    get:
      description: Returns the list of all users
      operationId: adminListUsers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Failed to list users
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List all users
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new user as an admin
      operationId: adminCreateUser
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Password must be at least 8 characters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to initialize user workspace
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create a new user
      tags:
      - Admin
  /admin/users/{userId}:
    delete:
      description: Delete a specific user as an admin
      operationId: adminDeleteUser
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Cannot delete your own account
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Cannot delete other admin users
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to delete user
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete a specific user
      tags:
      - Admin
    get:
      description: Get a specific user as an admin
      operationId: adminGetUser
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get a specific user
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update a specific user as an admin
      operationId: adminUpdateUser
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to update user
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update a specific user
      tags:
      - Admin
  /admin/workspaces:
    get:
      description: List all workspaces and their stats as an admin
      operationId: adminListWorkspaces
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.WorkspaceStats'
            type: array
        "500":
          description: Failed to get file stats
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List all workspaces
      tags:
      - Admin
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"